updating libraries
Ugh this is ugly.

lots of mods to the source to get the newer version of flow running,
using the flow tool documented on the flow blog to ignore issues.

deleted yarn.lock; switching everything to npm because yarn is not
running flow properly.

flow still isn't running quite right.  Have to manually run npm run
build steps.  Not sure what is the deal there, but so it goes when
you're cleaning up old code.

Also, got rid of usage of array-utils.  I started to clean that up
too, but it is just super annoying so I did not.  handful of uses in
this code were manually added.